requirements.txt
streamlit==1.40.2
pillow==10.4.0
